application commit 1) application product dialog to'g'irlandu 2) akt ni ichidagi documentlar tuzatildi 3) Queue tug'irlandi

+++ Domain/src/main/java/uz/einvoice/domain/models/documents/actDocument/Data.kt
@@ -6,7 +6,7 @@ data class Data(
     val buyerbranchcode: String,
     val buyerbranchname: String,
     val buyername: String,
-    val buyertin: Int,
+    val buyertin: Long,
     val contractdoc: Contractdoc,
     val covotingstate: Int,
     val created_at: String,
